Merge tables to a single table
In some cases, software will generate plenty of tab delimit text files under the same directory.
If you want to merge all the tables into one single table, I find a R package for this purpose. 
One thing to be sure is that the column should be exactly the same.